// src/app/models/data/group-result.ts
export interface GroupResultJson {
  id: number;

  groupGuid: string;
  groupName: string;
  groupId: string;
  subGroupGuid: string;
  subGroupName: string;
  subGroupId: string;
  address1: string;
  address2: string;
  city: string;
  state: string;
  termDate: string;
  futureEffectiveDate: string;
}